refactor: 컴포넌트 폴더 위치 변경
pages 폴더에는 page 관련 컴포넌트만 있게 변경
