+++ src/components/features/payment/PaymentCheckContents/PaymentCheckContents.styles.ts
