--- src/pages/cart/components/loadingContents/LoadingContents.styles.ts
@@ -1,10 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Container = styled.div`
-  flex: 1;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
