--- src/pages/cart/components/loadingContents/LoadingContents.tsx
@@ -1,12 +1,0 @@
-import { Loading } from '@/components/common';
-import * as S from './LoadingContents.styles';
-
-function LoadingContents() {
-  return (
-    <S.Container>
-      <Loading />
-    </S.Container>
-  );
-}
-
-export default LoadingContents;
