--- src/pages/notFound/NotFoundPage.styles.ts
@@ -1,26 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-  height: 100vh;
-  background-color: white;
-`;
-
-export const Title = styled.h1`
-  font-size: 24px;
-  font-weight: 700;
-  color: black;
-`;
-
-export const HomeButton = styled.button`
-  padding: 16px 20px;
-  font-size: 16px;
-  font-weight: 700;
-  background-color: black;
-  color: white;
-  border-radius: 4px;
-`;
