+++ src/pages/notFound/NotFoundPage.tsx
@@ -1,6 +1,6 @@
 import { ROUTE } from '@/shared';
 import { useNavigate } from 'react-router';
-import * as S from './NotFoundPage.styles';
+import styled from '@emotion/styled';
 
 function NotFoundPage() {
   const navigate = useNavigate();
@@ -10,11 +10,36 @@ function NotFoundPage() {
   };
 
   return (
-    <S.Container>
-      <S.Title>Page Not Found</S.Title>
-      <S.HomeButton onClick={handleGoHome}>메인 페이지로</S.HomeButton>
-    </S.Container>
+    <Container>
+      <Title>Page Not Found</Title>
+      <HomeButton onClick={handleGoHome}>메인 페이지로</HomeButton>
+    </Container>
   );
 }
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+  height: 100vh;
+  background-color: white;
+`;
+
+const Title = styled.h1`
+  font-size: 24px;
+  font-weight: 700;
+  color: black;
+`;
+
+const HomeButton = styled.button`
+  padding: 16px 20px;
+  font-size: 16px;
+  font-weight: 700;
+  background-color: black;
+  color: white;
+  border-radius: 4px;
+`;
 
 export default NotFoundPage;
